Move dynamic tool models to tool_util
